converting everything to guest_user instead of user in prep for user accounts
