+++ app/controllers/application_controller.rb
@@ -3,7 +3,7 @@ class ApplicationController < ActionController::Base
 
   def current_user
     @current_user ||= session[:current_user_id] &&
-      User.find_by(id: session[:current_user_id])
+      GuestUser.find_by(id: session[:current_user_id])
   end
 
 end
