+++ app/views/goals/introductions.html.erb
@@ -37,7 +37,7 @@ $(document).ready(function(){
 
 <div id="chat3" class="hidden" style="display:none;">
 
-<%= form_for @user do |f| %>
+<%= form_for @guest_user do |f| %>
 <section id="user-section">
 
 	<div class="spaced-row user-content">
