+++ app/views/guest_users/create.html.erb
@@ -1,0 +1,2 @@
+<h1>GuestUsers#create</h1>
+<p>Find me in app/views/guest_users/create.html.erb</p>
