+++ app/views/guest_users/new.html.erb
@@ -1,0 +1,2 @@
+<h1>GuestUsers#new</h1>
+<p>Find me in app/views/guest_users/new.html.erb</p>
