+++ app/views/intro/future.html.erb
@@ -33,7 +33,7 @@ $(document).ready(function(){
 </section>
 
 <div id= "chat2" class="hidden" style="display:none;">
-<%= form_for @user, url: wizard_path, method: :put do |f| %>
+<%= form_for @guest_user, url: wizard_path, method: :put do |f| %>
 	<%= f.fields_for :future_day do |i| %>
 <section id="user-section">
