+++ app/views/intro/savings.html.erb
@@ -39,7 +39,7 @@ $(document).ready(function(){
 
 <div id="chat2" class="hidden" style="display:none;">
 
-<%= form_for @user, url: wizard_path, method: :put do |f| %>
+<%= form_for @guest_user, url: wizard_path, method: :put do |f| %>
 
 	<section id="user-section">
