+++ app/views/intro/student_loans_info.html.erb
@@ -51,7 +51,7 @@ $(document).ready(function(){
 
 <div id="chat3" class="hidden" style="display:none;">
 	
-	<%= form_for @user, url: wizard_path, method: :put do |f| %>
+	<%= form_for @guest_user, url: wizard_path, method: :put do |f| %>
 
 	<section id="user-section">
