+++ app/views/intro/thanks.html.erb
@@ -10,7 +10,7 @@
 
 	<div class="coach-question">
 
-		Thanks, <%= @user.name %>. We'll be in touch soon!
+		Thanks, <%= @guest_user.name %>. We'll be in touch soon!
 	</div>
 </section>
 </body>
