--- db/migrate/20160802205311_add_email_to_user.rb
@@ -1,5 +1,0 @@
-class AddEmailToUser < ActiveRecord::Migration[5.0]
-  def change
-    add_column :users, :email, :string
-  end
-end
