--- db/migrate/20160804183928_change_investments_type_to_hstore.rb
@@ -1,5 +1,0 @@
-class ChangeInvestmentsTypeToHstore < ActiveRecord::Migration[5.0]
-  def change
-  	change_column :users, :investments_type, 'hstore USING CAST(investments_type AS hstore)'
-  end
-end
