--- db/migrate/20160805033742_change_future_day_to_hstore.rb
@@ -1,5 +1,0 @@
-class ChangeFutureDayToHstore < ActiveRecord::Migration[5.0]
-  def change
-  	change_column :users, :future_day, 'hstore USING CAST(future_day AS hstore)'
-  end
-end
