--- db/migrate/20160805215312_add_future_preparedness_to_users.rb
@@ -1,5 +1,0 @@
-class AddFuturePreparednessToUsers < ActiveRecord::Migration[5.0]
-  def change
-  	add_column :users, :preparedness, :string
-  end
-end
