--- db/migrate/20160805232719_add_savvy_score_components_to_users.rb
@@ -1,5 +1,0 @@
-class AddSavvyScoreComponentsToUsers < ActiveRecord::Migration[5.0]
-  def change
-  	add_column :users, :savvy_score_components, :hstore
-  end
-end
