--- db/migrate/20160808165933_add_financial_help_to_users.rb
@@ -1,5 +1,0 @@
-class AddFinancialHelpToUsers < ActiveRecord::Migration[5.0]
-  def change
-    add_column :users, :financial_help, :string
-  end
-end
