+++ test/fixtures/guest_users.yml
@@ -1,0 +1,23 @@
+# Read about fixtures at http://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  name: MyString
+  age: 1
+  student_amount: 9.99
+  cc_amount: 9.99
+  savings_amount: 9.99
+  investments_type: 
+  spend_vs_income: MyString
+  future_day: 
+  preparedness: MyString
+
+two:
+  name: MyString
+  age: 1
+  student_amount: 9.99
+  cc_amount: 9.99
+  savings_amount: 9.99
+  investments_type: 
+  spend_vs_income: MyString
+  future_day: 
+  preparedness: MyString
